chore(hybridcloud) Move outbox base classes
Move the outbox base classes into the hybridcloud module to help clarify
application design.
