+++ src/sentry/hybridcloud/outbox/base.py
