+++ src/sentry/integrations/models/external_actor.py
@@ -8,7 +8,7 @@ from django.utils import timezone
 from sentry.backup.scopes import RelocationScope
 from sentry.db.models import BoundedPositiveIntegerField, FlexibleForeignKey, region_silo_model
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
-from sentry.db.models.outboxes import ReplicatedRegionModel
+from sentry.hybridcloud.outbox.base import ReplicatedRegionModel
 from sentry.hybridcloud.services.replica import control_replica_service
 from sentry.integrations.types import ExternalProviders
 from sentry.models.outbox import OutboxCategory
