+++ src/sentry/models/apikey.py
@@ -11,7 +11,7 @@ from sentry.backup.scopes import RelocationScope
 from sentry.db.models import BoundedPositiveIntegerField, control_silo_model, sane_repr
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
 from sentry.db.models.manager.base import BaseManager
-from sentry.db.models.outboxes import ReplicatedControlModel
+from sentry.hybridcloud.outbox.base import ReplicatedControlModel
 from sentry.hybridcloud.services.replica import region_replica_service
 from sentry.models.apiscopes import HasApiScopes
 from sentry.models.outbox import OutboxCategory
