+++ src/sentry/models/authenticator.py
@@ -28,7 +28,7 @@ from sentry.db.models import (
 )
 from sentry.db.models.fields.picklefield import PickledObjectField
 from sentry.db.models.manager.base import BaseManager
-from sentry.db.models.outboxes import ControlOutboxProducingModel
+from sentry.hybridcloud.outbox.base import ControlOutboxProducingModel
 from sentry.models.outbox import ControlOutboxBase, OutboxCategory
 from sentry.types.region import find_regions_for_user
