+++ src/sentry/models/authidentity.py
@@ -12,7 +12,7 @@ from sentry.backup.sanitize import SanitizableField, Sanitizer
 from sentry.backup.scopes import RelocationScope
 from sentry.db.models import FlexibleForeignKey, control_silo_model, sane_repr
 from sentry.db.models.fields.jsonfield import JSONField
-from sentry.db.models.outboxes import ReplicatedControlModel
+from sentry.hybridcloud.outbox.base import ReplicatedControlModel
 from sentry.models.outbox import OutboxCategory
 from sentry.types.region import find_regions_for_orgs
