+++ src/sentry/models/authprovider.py
@@ -20,7 +20,7 @@ from sentry.db.models import (
 )
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
 from sentry.db.models.fields.jsonfield import JSONField
-from sentry.db.models.outboxes import ReplicatedControlModel
+from sentry.hybridcloud.outbox.base import ReplicatedControlModel
 from sentry.models.outbox import ControlOutbox, OutboxCategory, OutboxScope
 from sentry.types.region import find_regions_for_orgs
