+++ src/sentry/models/integrations/sentry_app_installation.py
@@ -14,8 +14,8 @@ from sentry.backup.scopes import RelocationScope
 from sentry.constants import SentryAppInstallationStatus
 from sentry.db.models import BoundedPositiveIntegerField, FlexibleForeignKey, control_silo_model
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
-from sentry.db.models.outboxes import ReplicatedControlModel
 from sentry.db.models.paranoia import ParanoidManager, ParanoidModel
+from sentry.hybridcloud.outbox.base import ReplicatedControlModel
 from sentry.projects.services.project import RpcProject
 from sentry.sentry_apps.services.app.model import RpcSentryAppComponent, RpcSentryAppInstallation
 from sentry.types.region import find_regions_for_orgs
