+++ src/sentry/models/integrations/sentry_app_installation_token.py
@@ -7,7 +7,7 @@ from django.db.models import QuerySet
 from sentry.backup.scopes import RelocationScope
 from sentry.db.models import FlexibleForeignKey, control_silo_model
 from sentry.db.models.manager.base import BaseManager
-from sentry.db.models.outboxes import ControlOutboxProducingModel
+from sentry.hybridcloud.outbox.base import ControlOutboxProducingModel
 from sentry.models.apitoken import ApiToken
 from sentry.models.outbox import ControlOutboxBase
