+++ src/sentry/models/organizationmemberteam.py
@@ -8,7 +8,7 @@ from django.db import models
 from sentry import features, roles
 from sentry.backup.scopes import RelocationScope
 from sentry.db.models import BoundedAutoField, FlexibleForeignKey, region_silo_model, sane_repr
-from sentry.db.models.outboxes import RegionOutboxProducingManager, ReplicatedRegionModel
+from sentry.hybridcloud.outbox.base import RegionOutboxProducingManager, ReplicatedRegionModel
 from sentry.models.outbox import OutboxCategory, RegionOutboxBase
 from sentry.roles import team_roles
 from sentry.roles.manager import TeamRole
