+++ src/sentry/models/orgauthtoken.py
@@ -16,7 +16,7 @@ from sentry.conf.server import SENTRY_SCOPES
 from sentry.db.models import ArrayField, FlexibleForeignKey, control_silo_model, sane_repr
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
 from sentry.db.models.manager.base import BaseManager
-from sentry.db.models.outboxes import ReplicatedControlModel
+from sentry.hybridcloud.outbox.base import ReplicatedControlModel
 from sentry.models.organization import Organization
 from sentry.models.outbox import OutboxCategory
 from sentry.utils.hashlib import sha1_text
