+++ src/sentry/models/team.py
@@ -20,8 +20,8 @@ from sentry.db.models import (
 )
 from sentry.db.models.fields.slug import SentrySlugField
 from sentry.db.models.manager.base import BaseManager
-from sentry.db.models.outboxes import ReplicatedRegionModel
 from sentry.db.models.utils import slugify_instance
+from sentry.hybridcloud.outbox.base import ReplicatedRegionModel
 from sentry.locks import locks
 from sentry.models.outbox import OutboxCategory
 from sentry.utils.retries import TimedRetryPolicy
