+++ src/sentry/models/useremail.py
@@ -21,7 +21,7 @@ from sentry.backup.sanitize import SanitizableField, Sanitizer
 from sentry.backup.scopes import ImportScope, RelocationScope
 from sentry.db.models import FlexibleForeignKey, control_silo_model, sane_repr
 from sentry.db.models.manager.base import BaseManager
-from sentry.db.models.outboxes import ControlOutboxProducingModel
+from sentry.hybridcloud.outbox.base import ControlOutboxProducingModel
 from sentry.models.outbox import ControlOutboxBase, OutboxCategory
 from sentry.organizations.services.organization.model import RpcOrganization
 from sentry.types.region import find_regions_for_user
