+++ src/sentry/models/userrole.py
@@ -10,7 +10,7 @@ from sentry.backup.mixins import OverwritableConfigMixin
 from sentry.backup.scopes import RelocationScope
 from sentry.db.models import ArrayField, control_silo_model, sane_repr
 from sentry.db.models.fields.foreignkey import FlexibleForeignKey
-from sentry.db.models.outboxes import ControlOutboxProducingModel
+from sentry.hybridcloud.outbox.base import ControlOutboxProducingModel
 from sentry.models.outbox import ControlOutboxBase, OutboxCategory
 from sentry.signals import post_upgrade
 from sentry.silo.base import SiloMode
