+++ src/sentry/tasks/backfill_outboxes.py
@@ -12,7 +12,7 @@ from django.db import router, transaction
 from django.db.models import Max, Min, Model
 
 from sentry import options
-from sentry.db.models.outboxes import ControlOutboxProducingModel, RegionOutboxProducingModel
+from sentry.hybridcloud.outbox.base import ControlOutboxProducingModel, RegionOutboxProducingModel
 from sentry.models.outbox import outbox_context
 from sentry.models.user import User
 from sentry.silo.base import SiloMode
